+++ website/src/components/HomePage/Tagline.js
@@ -4,8 +4,8 @@ function Tagline() {
   return (
     <section className="bg-gray-50 px-4 py-28">
       <h1 className="py-3 text-center text-4xl font-bold">Azure Web PubSub</h1>
-      <p className="text-center">Easily add real-time capabilities to your apps using your preferred tech stack.</p>
-      <p className="text-center font-semibold">Reliable. Secure. Highly scalable.</p>
+      <h2 className="text-center">Easily add real-time capabilities to your apps using your preferred tech stack.</h2>
+      <h2 className="text-center font-semibold">Reliable. Secure. Highly scalable.</h2>
     </section>
   )
 }